refactor: clean architecture (Core/Application/Infrastructure/Presentation)
- move Models -> Core/Entities (AppTask, CalendarEvent, KanbanColumn, etc.)
- move Data -> Core/Interfaces + Infrastructure/Repositories
- move Services -> Application/Services
- move Components/* -> Presentation/*
- update MauiProgram.cs namespaces (Jamrah.Core.Interfaces etc.)
- update MainPage.xaml.cs component types to Presentation.*
- update .gitignore: ignore docs/, .opencode/, tasks-page.html, *.html.mock
