+++ Application/Services/PomodoroSoundService.cs
@@ -6,7 +6,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.JSInterop;
 
-namespace Jamrah.Services
+namespace Jamrah.Application.Services
 {
     public enum PomodoroSoundEvent { SessionStart, TimerWarning, SessionComplete, BreakStart, BreakComplete }
