+++ Presentation/_Imports.razor
@@ -7,6 +7,6 @@
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Jamrah
-@using Jamrah.Models
-@using Jamrah.Services
-@using Jamrah.Data
+@using Jamrah.Core.Entities
+@using Jamrah.Core.Interfaces
+@using Jamrah.Application.Services
